# beta-private/api/NHSD.GPITF.BuyingCatalog/crm/Gif.Service/wwwroot/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0-private-beta",
    "title" : "catalogue-api",
    "description" : "NHS Digital GP IT Futures Buying Catalog API"
  },
  "paths" : {
...
    "/api/TechnicalContacts/BySolution/{solutionId}" : {
      "get" : {
        "tags" : [ "TechnicalContacts" ],
        "summary" : "Retrieve all Technical Contacts for a solution in a paged list,\r\ngiven the solution’s CRM identifier",
        "operationId" : "ApiTechnicalContactsBySolutionBySolutionIdGet",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solutionId",
          "in" : "path",
          "description" : "CRM identifier of solution",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "1-based index of page to return.  Defaults to 1",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "number of items per page.  Defaults to 20",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/PaginatedList[TechnicalContacts]"
            }
          },
          "404" : {
            "description" : "Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/api/TechnicalContacts" : {
      "put" : {
        "tags" : [ "TechnicalContacts" ],
        "summary" : "Update a Technical Contact with new information",
        "operationId" : "ApiTechnicalContactsPut",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "techCont",
          "in" : "body",
          "description" : "Technical Contact with updated information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TechnicalContacts"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "Technical Contact or Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "TechnicalContacts" ],
        "summary" : "Create a new Technical Contact for a Solution",
        "operationId" : "ApiTechnicalContactsPost",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "techCont",
          "in" : "body",
          "description" : "new Technical Contact information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TechnicalContacts"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/TechnicalContacts"
            }
          },
          "404" : {
            "description" : "Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "TechnicalContacts" ],
        "summary" : "Delete an existing Technical Contact for a Solution",
        "operationId" : "ApiTechnicalContactsDelete",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "techCont",
          "in" : "body",
          "description" : "existing Technical Contact information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TechnicalContacts"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "Technical Contact or Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    }
  },
  "definitions" : {
...
    "TechnicalContacts" : {
      "description" : "A means of communicating with an Organisation, typically a person, email address, telephone number etc,\r\nin the context of a Solution",
      "required" : [ "id", "solutionId", "contactType", "emailAddress" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "description" : "Unique identifier of entity",
          "type" : "string"
        },
        "solutionId" : {
          "description" : "Unique identifier of Solution",
          "type" : "string"
        },
        "contactType" : {
          "description" : "Description of type of TechnicalContact eg\r\n<list type=\"bullet\">\r\nLead Contact\r\nTechnical Contact\r\nExecutive Sponsor\r\nClinical Safety Officer\r\nConnection Agreement Signatory\r\nOther...\r\n</list>",
          "type" : "string"
        },
        "firstName" : {
          "description" : "First name of contact",
          "type" : "string"
        },
        "lastName" : {
          "description" : "Last name of contact",
          "type" : "string"
        },
        "emailAddress" : {
          "description" : "Primary email address of contact",
          "type" : "string"
        },
        "phoneNumber" : {
          "description" : "Primary phone number of contact",
          "type" : "string"
        }
      },
      "example" : {
        "id" : "87a58dbf-ab78-4350-9f3d-9276de3e4be6",
        "solutionId" : "A3C6830F-2E7C-4545-A4B9-02D20C4C92E1",
        "contactType" : "Lead Contact",
        "firstName" : "Jon",
        "lastName" : "Dough",
        "emailAddress" : "jon.dough @tpp.com",
        "phoneNumber" : "0300 303 5678"
      }
    },
...
    "PaginatedList[TechnicalContacts]" : {
      "description" : "A paged list of objects",
      "type" : "object",
      "properties" : {
        "pageIndex" : {
          "format" : "int32",
          "description" : "1-based index of which page this page\r\nDefaults to 1",
          "type" : "integer",
          "readOnly" : true
        },
        "totalPages" : {
          "format" : "int32",
          "description" : "Total number of pages based on NHSD.GPITF.BuyingCatalog.Models.PaginatedList`1.PageSize",
          "type" : "integer",
          "readOnly" : true
        },
        "pageSize" : {
          "format" : "int32",
          "description" : "Maximum number of items in this page\r\nDefaults to 20",
          "type" : "integer",
          "readOnly" : true
        },
        "items" : {
          "description" : "List of items",
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TechnicalContacts"
          },
          "readOnly" : true
        }
      }
    }
  },
  "securityDefinitions" : {
    "Bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  }
}